#214: rename search 'navigation' to 'coordintor' and make parent element
