+++ templates/search_pages/index.html
@@ -18,25 +18,20 @@
   }
 </style>
 
-<vivo-i18n-labels id="search-terms">
-  <vivo-i18n-label key="show-more" label="<%= t("show_more") %>" />
-  <vivo-i18n-label key="show-less" label="<%= t("show_less") %>" />
-  <vivo-i18n-label key="cancel" label="<%= t("cancel") %>" />
-  <vivo-i18n-label key="apply" label="<%= t("apply") %>" />
-  <!-- pagination -->
-  <vivo-i18n-label key="next" label="<%= t("next") %>" />
-  <vivo-i18n-label key="previous" label="<%= t("previous") %>" />
-  <vivo-i18n-label key="showing" label="<%= t("showing") %>" />
-  <vivo-i18n-label key="of" label="<%= t("of") %>" />
-</vivo-i18n-labels>
-
-<vivo-search-navigation labels="search-terms">
-</vivo-search-navigation>
-
 <vivo-modal id="search-waiting">
   <h3><%= t("searching") %>...</h3>
   <vivo-search-spinner></vivo-search-spinner>
 </vivo-modal>
+
+<vivo-search-coordinator>
+  <vivo-i18n-label key="show-more" label="<%= t("show_more") %>"></vivo-i18n-label>
+  <vivo-i18n-label key="show-less" label="<%= t("show_less") %>"></vivo-i18n-label>
+  <vivo-i18n-label key="cancel" label="<%= t("cancel") %>"></vivo-i18n-label>
+  <vivo-i18n-label key="apply" label="<%= t("apply") %>"></vivo-i18n-label>
+  <vivo-i18n-label key="next" label="<%= t("next") %>"></vivo-i18n-label>
+  <vivo-i18n-label key="previous" label="<%= t("previous") %>"></vivo-i18n-label>
+  <vivo-i18n-label key="showing" label="<%= t("showing") %>"></vivo-i18n-label>
+  <vivo-i18n-label key="of" label="<%= t("of") %>"></vivo-i18n-label>
 
 <div class="search-container">
   <!-- NOTE: facets and search are in separate sections
@@ -182,6 +177,7 @@
               <vivo-search-sort-option label="<%= t("title_ascending") %>" field="title" direction="ASC" default="true"/>
               <vivo-search-sort-option label="<%= t("title_descending") %>" field="title" direction="DESC" />
             </vivo-search-sort-options>
+            <!-- TODO: here or just in component? -->
             <vivo-search-pagination slot="pagination">
             </vivo-search-pagination>
             <vivo-search-pagination-summary slot="pagination-summary">
@@ -200,6 +196,8 @@
     </vivo-tabs>
 
   </main>
-</div>
+</div><!-- end .search-container -->
+
+</vivo-search-coordinator>
 
 <%= javascriptTag("search.js") %>
